# app/go/hospital-app/appointments/kodata/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Appointments Service
  description: Appointments Service
  version: 1.0.0
  contact:
    email: salaboy@gmail.com
    name: Mauricio Salatino
    url: https://salaboy.com
servers:
  - url: http://localhost:8080
paths:
  /appointments/:
    post:
      summary: Create Appointment
      operationId: createAppointment
      tags:
        - appointments
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Appointment"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/Appointment"
    get:
      summary: Get All Appointments
      operationId: getAllAppointments
      tags:
        - appointments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
  /service/info:
    get:
      summary: Get Service Info
      operationId: getServiceInfo
      tags:
        - service
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                items:
                  $ref: "#/components/schemas/ServiceInfo"
components:
  schemas:
    Appointment:
      properties:
        id:
          type: string
          description: Appointment ID
        patientId:
          type: string
          description: Patient ID
        deparmentId:
          type: string
          description: Department ID
        title:
          type: string
          description: Notification Title
    ServiceInfo:
      properties:
        name:
          type: string
          description: Service Name
        version:
          type: string
          description: Service Version
        source:
          type: string
          description: Service Source
        podId:
          type: string
          description: Service Pod ID
        podNamespace:
          type: string
          description: Service Pod Namespace
        podNodeName:
          type: string
          description: Service Pod Node Name